requirements.txt
librosa==0.10.1
SpeechRecognition==3.10.0
numpy>=1.23.5
pydub==0.25.1
ffmpeg-python>=0.2.0
mediapipe>=0.10.0
opencv-python>=4.7.0
google-generativeai>=0.3.0
tensorflow==2.12.0
requests>=2.28.2
tk>=0.1.0
chromadb>=0.6.0
tqdm>=4.65.0 